Fix dtype issues in array indexing
Explicitly cast index arrays to np.int64 to avoid IndexError when
numpy infers object dtype from list comprehensions. Affects:
- nuisance_estimation.py: unit_rows for sample ID lookup
- aggregation.py: cluster_indices for event study aggregation
- inference.py: cluster_indices for bootstrap

+++ code/python/modules/aggregation.py
@@ -220,8 +220,8 @@ def aggregate_event_study(att_results: Dict, config: Config) -> Dict:
     n_clusters = len(clusters)
     cluster_map = pd.Series(range(len(clusters)), index=clusters)
 
-    # Map unit IDs to cluster indices
-    cluster_indices = np.array([cluster_map[unit_clusters[uid]] for uid in unit_ids])
+    # Map unit IDs to cluster indices (explicit int64 for indexing)
+    cluster_indices = np.array([cluster_map[unit_clusters[uid]] for uid in unit_ids], dtype=np.int64)
 
     log_message(f"  Computing clustered bootstrap ({config.inference.n_bootstrap} reps, {n_clusters} clusters)...")
